+++ tests/test_csc.py
@@ -22,9 +22,16 @@ import asyncio
 import pathlib
 import unittest
 
+import pytest
+
 from lsst.ts import salobj
 from lsst.ts import hexrotcomm
-from lsst.ts.idl.enums.MTRotator import ControllerState
+from lsst.ts.idl.enums.MTRotator import (
+    ApplicationStatus,
+    ControllerState,
+    EnabledSubstate,
+    ErrorCode,
+)
 
 STD_TIMEOUT = 5  # timeout for command ack
 
@@ -32,9 +39,7 @@ LOCAL_CONFIG_DIR = pathlib.Path(__file__).parent / "data" / "config"
 
 
 class TestSimpleCsc(hexrotcomm.BaseCscTestCase, unittest.IsolatedAsyncioTestCase):
-    def basic_make_csc(
-        self, config_dir=None, initial_state=salobj.State.OFFLINE, simulation_mode=1
-    ):
+    def basic_make_csc(self, config_dir, initial_state, simulation_mode):
         return hexrotcomm.SimpleCsc(
             initial_state=initial_state,
             simulation_mode=simulation_mode,
@@ -42,28 +47,29 @@ class TestSimpleCsc(hexrotcomm.BaseCscTestCase, unittest.IsolatedAsyncioTestCase
         )
 
     async def test_constructor_errors(self):
-        for bad_initial_state in (0, max(salobj.State) + 1):
-            with self.assertRaises(ValueError):
+        for bad_initial_state in (0, salobj.State.OFFLINE, max(salobj.State) + 1):
+            with pytest.raises(ValueError):
                 hexrotcomm.SimpleCsc(initial_state=bad_initial_state, simulation_mode=1)
 
-        for bad_simulation_mode in (-1, 2):
-            with self.assertRaises(ValueError):
+        for bad_simulation_mode in (-1, 0, 2):
+            with pytest.raises(ValueError):
                 hexrotcomm.SimpleCsc(
-                    initial_state=bad_initial_state,
+                    initial_state=salobj.State.STANDBY,
                     simulation_mode=bad_simulation_mode,
                 )
 
-        with self.assertRaises(ValueError):
+        with pytest.raises(ValueError):
             hexrotcomm.SimpleCsc(
-                initial_state=salobj.State.OFFLINE,
+                initial_state=salobj.State.STANDBY,
                 simulation_mode=1,
                 config_dir="no_such_directory",
             )
 
+        # When not simulating the only valid initial state is STANDBY
         for bad_initial_state in salobj.State:
-            if bad_initial_state == salobj.State.OFFLINE:
+            if bad_initial_state == salobj.State.STANDBY:
                 continue
-            with self.assertRaises(ValueError):
+            with pytest.raises(ValueError):
                 hexrotcomm.SimpleCsc(
                     initial_state=bad_initial_state,
                     simulation_mode=0,
@@ -75,20 +81,117 @@ class TestSimpleCsc(hexrotcomm.BaseCscTestCase, unittest.IsolatedAsyncioTestCase
             simulation_mode=1,
             config_dir=LOCAL_CONFIG_DIR,
         ):
-
-            # Try a config file that has invalid data.
+            # Try config files with invalid data.
             # The command should fail and the summary state remain in STANDBY.
-            with salobj.assertRaisesAckError():
-                await self.remote.cmd_start.set_start(
-                    settingsToApply="invalid.yaml", timeout=STD_TIMEOUT
-                )
-            self.assertEqual(self.csc.summary_state, salobj.State.STANDBY)
+            for bad_config_path in LOCAL_CONFIG_DIR.glob("bad_*.yaml"):
+                bad_config_name = bad_config_path.name
+                with self.subTest(bad_config_name=bad_config_name):
+                    with salobj.assertRaisesAckError():
+                        await self.remote.cmd_start.set_start(
+                            settingsToApply=bad_config_name, timeout=STD_TIMEOUT
+                        )
+                    assert self.csc.summary_state == salobj.State.STANDBY
 
             # Now try a valid config file
             await self.remote.cmd_start.set_start(
                 settingsToApply="valid.yaml", timeout=STD_TIMEOUT
             )
-            self.assertEqual(self.csc.summary_state, salobj.State.DISABLED)
+            assert self.csc.summary_state == salobj.State.DISABLED
+
+    async def test_controller_fault(self):
+        """Controller going to fault should send CSC to fault, error code 1"""
+        async with self.make_csc(
+            initial_state=salobj.State.ENABLED,
+            simulation_mode=1,
+            config_dir=LOCAL_CONFIG_DIR,
+        ):
+            await self.assert_next_sample(
+                topic=self.remote.evt_controllerState,
+                controllerState=ControllerState.ENABLED,
+                enabledSubstate=EnabledSubstate.STATIONARY,
+            )
+            await self.assert_next_summary_state(salobj.State.ENABLED)
+
+            self.csc.mock_ctrl.set_state(ControllerState.FAULT)
+            await self.assert_next_sample(
+                topic=self.remote.evt_controllerState,
+                controllerState=ControllerState.FAULT,
+            )
+            await self.assert_next_summary_state(salobj.State.FAULT)
+            await self.assert_next_sample(
+                topic=self.remote.evt_errorCode, errorCode=ErrorCode.CONTROLLER_FAULT
+            )
+
+    async def test_lose_connection(self):
+        """Losing the connection should send CSC to fault, error code 2"""
+        async with self.make_csc(
+            initial_state=salobj.State.ENABLED,
+            simulation_mode=1,
+            config_dir=LOCAL_CONFIG_DIR,
+        ):
+            assert self.csc.client.should_be_connected
+            await self.assert_next_sample(
+                topic=self.remote.evt_controllerState,
+                controllerState=ControllerState.ENABLED,
+                enabledSubstate=EnabledSubstate.STATIONARY,
+            )
+            await self.assert_next_summary_state(salobj.State.ENABLED)
+
+            assert self.csc.client.should_be_connected
+            await self.csc.mock_ctrl.close_client()
+            await self.assert_next_summary_state(salobj.State.FAULT)
+            await self.assert_next_sample(
+                topic=self.remote.evt_errorCode, errorCode=ErrorCode.CONNECTION_LOST
+            )
+
+    async def test_controller_not_enabled(self):
+        """Controller going out of enabled state should disable the CSC"""
+        async with self.make_csc(
+            initial_state=salobj.State.ENABLED,
+            simulation_mode=1,
+            config_dir=LOCAL_CONFIG_DIR,
+        ):
+            await self.assert_next_sample(
+                topic=self.remote.evt_controllerState,
+                controllerState=ControllerState.ENABLED,
+                enabledSubstate=EnabledSubstate.STATIONARY,
+            )
+            await self.assert_next_summary_state(salobj.State.ENABLED)
+
+            self.csc.mock_ctrl.set_state(ControllerState.DISABLED)
+            await self.assert_next_sample(
+                topic=self.remote.evt_controllerState,
+                controllerState=ControllerState.DISABLED,
+            )
+            await self.assert_next_summary_state(salobj.State.DISABLED)
+
+    async def test_eui_takes_control(self):
+        """If the EUI takes control this should disable the CSC"""
+        async with self.make_csc(
+            initial_state=salobj.State.ENABLED,
+            simulation_mode=1,
+            config_dir=LOCAL_CONFIG_DIR,
+        ):
+            await self.assert_next_sample(
+                topic=self.remote.evt_controllerState,
+                controllerState=ControllerState.ENABLED,
+                enabledSubstate=EnabledSubstate.STATIONARY,
+            )
+            await self.assert_next_summary_state(salobj.State.ENABLED)
+            await self.assert_next_sample(
+                topic=self.remote.evt_commandableByDDS,
+                state=True,
+            )
+
+            # Clear the DDS_COMMAND_SOURCE flag
+            self.csc.mock_ctrl.telemetry.application_status &= (
+                ~ApplicationStatus.DDS_COMMAND_SOURCE
+            )
+            await self.assert_next_sample(
+                topic=self.remote.evt_commandableByDDS,
+                state=False,
+            )
+            await self.assert_next_summary_state(salobj.State.DISABLED)
 
     async def move_sequentially(self, *positions, delay=None):
         """Move sequentially to different positions, in order to test
@@ -123,12 +226,12 @@ class TestSimpleCsc(hexrotcomm.BaseCscTestCase, unittest.IsolatedAsyncioTestCase
                 controllerState=ControllerState.ENABLED,
             )
             data = await self.remote.tel_rotation.next(flush=True, timeout=STD_TIMEOUT)
-            self.assertAlmostEqual(data.demandPosition, 0)
+            assert data.demandPosition == pytest.approx(0)
             await self.remote.cmd_move.set_start(
                 position=destination, timeout=STD_TIMEOUT
             )
             data = await self.remote.tel_rotation.next(flush=True, timeout=STD_TIMEOUT)
-            self.assertAlmostEqual(data.demandPosition, destination)
+            assert data.demandPosition == pytest.approx(destination)
 
     async def test_run_multiple_commands(self):
         """Test BaseCsc.run_multiple_commands."""
@@ -166,22 +269,14 @@ class TestSimpleCsc(hexrotcomm.BaseCscTestCase, unittest.IsolatedAsyncioTestCase
             await self.csc.do_move(other_move)
 
             # task1 should have finished before the do_move command.
-            self.assertTrue(task1.done())
+            assert task1.done()
 
             # Wait for final telemetry.
             await asyncio.sleep(telemetry_delay)
 
             expected_positions = [0] + list(target_positions) + [other_move.position]
-            self.assertEqual(expected_positions, demand_positions)
+            assert expected_positions == demand_positions
 
     async def test_standard_state_transitions(self):
         async with self.make_csc(initial_state=salobj.State.STANDBY, simulation_mode=1):
             await self.check_standard_state_transitions(enabled_commands=("move",))
-
-            # Check that the fault method is not implemented
-            with self.assertRaises(NotImplementedError):
-                self.csc.fault(code=1, report="this should raise NotImplementedError")
-
-
-if __name__ == "__main__":
-    unittest.main()